+++ src/main/kotlin/org/jetbrains/changelog/extensions.kt
@@ -3,6 +3,8 @@ package org.jetbrains.changelog
 import groovy.lang.Closure
 import org.intellij.markdown.html.HtmlGenerator
 import org.intellij.markdown.parser.MarkdownParser
+import org.jetbrains.changelog.flavours.ChangelogFlavourDescriptor
+import org.jetbrains.changelog.flavours.PlainTextFlavourDescriptor
 
 fun <T : Any> closure(function: () -> T) = object : Closure<T>(null) {
     @Suppress("unused")
@@ -10,5 +12,11 @@ fun <T : Any> closure(function: () -> T) = object : Closure<T>(null) {
 }
 
 fun markdownToHTML(input: String) = ChangelogFlavourDescriptor().run {
-    HtmlGenerator(input, MarkdownParser(this).buildMarkdownTreeFromString(input), this, false).generateHtml()
+    HtmlGenerator(input, MarkdownParser(this).buildMarkdownTreeFromString(input), this, false)
+        .generateHtml()
+}
+
+fun markdownToPlainText(input: String) = PlainTextFlavourDescriptor().run {
+    HtmlGenerator(input, MarkdownParser(this).buildMarkdownTreeFromString(input), this, false)
+        .generateHtml(PlainTextTagRenderer())
 }