+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.3.2]
 ### Added
 - `markdownToHTML` method in `extensions.kt` file
+- `markdownToPlainText` method in `extensions.kt` file
 
 ## [0.3.1]
 ### Added